feat(terraform): change function app replacement
Currently, when the python code changes, the azure
function is not updated because the terraform resource
"azurerm_linux_function_app" doesn't have natively a way
to update itself.
Previously, I used a trick with the function name to avoid
this problem, the function name was defined with a prefix and
the python code hash. Unfortunately. with this trick redeployments
were too long because almost all the stack was replaced.

This commit adds the terraform lifecycle "replace_triggered_by" to
trigger the function app redeployment each time the Python code
changes instead of using a tricky hack with the app function name.